Remove useless wave function storages in k_point

+++ src/input.h
@@ -602,7 +602,7 @@ struct Hubbard_input
             }
 
             // now convert eV in Ha
-            for (int s = 0; s < coef_.size() - 3; s++) {
+            for (int s = 0; s < static_cast<int>(coef_.size() - 3); s++) {
                 coef_[s] *= 0.0367493;
             }
